@@ -5,7 +5,7 @@ author: "Ming Chen <ming_c@hotmail.com>"
 homepage: https://github.com/mingchen/secure_random
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
 
