@@ -1,5 +1,3 @@
-library secure_random;
-
 import 'dart:math';
 
 /// Secure random generator for string / number.
@@ -7,7 +5,7 @@ import 'dart:math';
 /// This class use cryptographically secure source of random numbers.
 ///
 class SecureRandom {
-  Random _random;
+  late Random _random;
   static const String _defaultCharset =
       'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
 
@@ -16,15 +14,7 @@ class SecureRandom {
     _random = Random.secure();
   }
 
-  /// Generate a strong random string.
-  ///
-  /// @param length   The length of target random string.
-  /// @param charset  Characters used in random string.
-  ///                 The default `charset` is 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
-  ///
-  /// @returns Random string with a fixed length.
-  ///
-  String nextString({int length, String charset = _defaultCharset}) {
+  String nextString({required int length, String charset = _defaultCharset}) {
     String ret = '';
 
     for (var i = 0; i < length; ++i) {
@@ -36,7 +26,7 @@ class SecureRandom {
   }
 
   /// Generate a strong random integer.
-  int nextInt({int max}) {
+  int nextInt({required int max}) {
     return _random.nextInt(max);
   }
 }
